Fix type instability in LinearSolve QR factorization fallback
Requires https://github.com/SciML/LinearSolve.jl/pull/617

This now uses the DefaultLinearSolver sum type in order to do the QR factorization fallback. This removes the `additional_lincache` which was `::Any`. This fixes a core type instability on the `LinearSolveResult` that was causing a pretty decently large slow down to the Newton methods. This also makes it so if the user specifically says they want LU, it don't do anything different.