Remove debugging macros in `reinit!` due to overhead
This ends up having a lot of overhead found in https://github.com/SciML/OrdinaryDiffEq.jl/issues/2757. They are triggered because the `reinit!` code recursively `reinit!`s the internals of the cache, and `LinearCache` doesn't have a `reinit!`, so it's a no-op. This is correct, but it turns out the debugging macro is just really expensive so it makes it slow even though it's not doing anything 😅 . I commented it out so we know where to put SciMLVerbosity.jl instructions for the near future @jClugstor